lms-system/backend/requirements/development.txt
-r base.txt

# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Git hooks
pre-commit==3.6.1

# Debug
django-debug-toolbar==4.3.0